+++ src/Services/FinanceService/FinanceService.Application/Contracts/CurrencyRateDto.cs
@@ -1,3 +1,3 @@
 namespace FinanceService.Application.Contracts;
 
-public sealed record CurrencyRateDto(string Name, decimal Rate);
+public sealed record CurrencyRateDto(string Name, double Rate);
